requirements.txt


amqp==5.2.0
argcomplete==3.5.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.50
botocore==1.35.50
bs4==0.0.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfn-flip==1.3.0
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
durationpy==0.9
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
kappa==0.6.0
kombu==5.4.2
MarkupSafe==3.0.2
packaging==24.2
pillow==10.4.0
placebo==0.9.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
service-identity==24.2.0
setuptools==75.2.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.6
troposphere==4.8.3
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
zappa==0.59.0
zope.interface==7.1.1
